+++ src/SharedAssemblyInfo.fs
@@ -4,10 +4,10 @@ open System.Reflection
 [<assembly: AssemblyCompanyAttribute("Yaaf.FSharp.Scripting")>]
 [<assembly: AssemblyProductAttribute("Yaaf.FSharp.Scripting")>]
 [<assembly: AssemblyCopyrightAttribute("Yaaf.FSharp.Scripting Copyright © Matthias Dittrich 2015")>]
-[<assembly: AssemblyVersionAttribute("1.1.1")>]
-[<assembly: AssemblyFileVersionAttribute("1.1.1")>]
-[<assembly: AssemblyInformationalVersionAttribute("1.1.1")>]
+[<assembly: AssemblyVersionAttribute("1.1.2")>]
+[<assembly: AssemblyFileVersionAttribute("1.1.2")>]
+[<assembly: AssemblyInformationalVersionAttribute("1.1.2")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "1.1.1"
+    let [<Literal>] Version = "1.1.2"